+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-04-02 21:59:46
-// Commit: refactor(test): migrate test controller to modern notifier syntax
+// Updated at: 2026-04-04 14:17:35
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '88';
-  final String lastMessage = 'refactor(test): migrate test controller to modern notifier syntax';
-  final DateTime timestamp = DateTime.parse('2026-04-02T21:59:46+05:00');
+  final String commitRef = '111';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-04-04T14:17:35+05:00');
 }